fdsf
- Trying to redo context tog et it to uåload properly
