+++ src/App.tsx
@@ -2,16 +2,18 @@ import { Loader } from "@navikt/ds-react";
 import { ErrorBoundary } from "react-error-boundary";
 import { RouterProvider } from "react-router-dom";
 import { router } from "./routes/Routes";
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import sanityClient from "./client";
 import { ProjectContext } from "./context/ProjectContext";
-import { UserContext, initialUserState } from "./context/UserContext";
+import { UserContext, initialUserState } from "./context/UserProvider";
 import { IProject } from "./types/project";
 import { IUser } from "./types/user";
 
 function App() {
   const [projects, setProjects] = useState<IProject[]>([]);
   const [user, setUser] = useState<IUser>(initialUserState);
+
+  const userValue = useMemo(() => ({ user, setUser }), [user]);
 
   useEffect(() => {
     sanityClient
@@ -50,7 +52,7 @@ function App() {
 
   return (
     <ErrorBoundary fallback={<div>Something went wrong!</div>}>
-      <UserContext.Provider value={user ?? {}}>
+      <UserContext.Provider value={userValue ?? {}}>
         <ProjectContext.Provider value={projects ?? []}>
           <RouterProvider router={router} fallbackElement={<Loader />} />
         </ProjectContext.Provider>
